Fix Vulkan animation jitter: double-buffer dynamic VBO + timing fixes
The primary bug was a CPU/GPU race on the SpriteBatch dynamic vertex
buffer. With FRAMES_IN_FLIGHT=2, the CPU was overwriting the single
host-visible VBO on frame N while the GPU was still reading it for
frame N-1, corrupting vertex positions and UV coordinates each frame.

Fix: VulkanVertexBuffer (dynamic mode) now allocates FRAMES_IN_FLIGHT
host-visible buffers. update() and bufferHandle() index by currentFrame,
which is safe because VulkanContext.clear() already waits on the
in-flight fence for that slot before any CPU writes happen.

Also fixed:
- Time.timeStarted stored as float caused ~1-2ms delta time quantization
  (float lacks precision at typical nanoTime magnitudes); changed to long
  so the subtraction is exact before narrowing to float.
- VulkanSwapChain uses minImageCount instead of minImageCount+1 to reduce
  frame pipelining depth and improve vsync pacing.

+++ core/src/main/java/hu/mudlee/core/Time.java
@@ -1,7 +1,7 @@
 package hu.mudlee.core;
 
 public class Time {
-    public static float timeStarted = System.nanoTime();
+    public static long timeStarted = System.nanoTime();
 
     public static float getTime() {
         return (float) ((System.nanoTime() - timeStarted) * 1E-9);
